[UPDATE] 리팩토링
- 기존 AuthorizationPlugin 플로그인 삭제
- AuthRouting 삭제
- DatabaseFactory 수정
  - connect() 함수 수정
  - dbQuery() 함수 생성
- '/users' 라우트 수정
  - '/login' 라우트 분리
