--- src/main/kotlin/com/sslmo/api/v1/auth/AuthRouting.kt
@@ -1,13 +1,0 @@
-package com.sslmo.api.v1.auth
-
-import io.ktor.server.routing.*
-
-fun Route.authRouting() {
-
-
-    route("/auth") {
-//        signIn()
-//        signCheck()
-
-    }
-}
