+++ src/main/kotlin/com/sslmo/database/tables/UserTable.kt
@@ -1,9 +1,8 @@
-package com.sslmo.database
+package com.sslmo.database.tables
 
+import com.sslmo.database.sql_types.uuid
 import com.sslmo.models.SignType
-import kotlinx.serialization.Contextual
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.Transient
+import com.sslmo.models.user.User
 import org.ktorm.database.Database
 import org.ktorm.dsl.QueryRowSet
 import org.ktorm.entity.sequenceOf
@@ -11,32 +10,6 @@ import org.ktorm.schema.*
 import java.time.LocalDate
 import java.util.*
 
-
-@Serializable
-data class User(
-    val id: Int,
-
-    @Contextual
-    val uuid: UUID,
-    val type: SignType,
-    val socialId: String?,
-    val email: String,
-
-    @Transient
-    val password: String? = null,
-    val nickname: String,
-    val notification: Boolean,
-    val active: Boolean,
-
-    @Contextual
-    val inActiveAt: LocalDate?,
-
-    @Contextual
-    val createdAt: LocalDate,
-
-    @Contextual
-    val updatedAt: LocalDate?,
-)
 
 object Users : BaseTable<User>("user") {
     val id = int("id").primaryKey()
