+++ src/main/kotlin/com/sslmo/models/user/LoginResponse.kt
@@ -1,6 +1,5 @@
 package com.sslmo.models.user
 
-import com.sslmo.database.User
 import com.sslmo.utils.Token
 import kotlinx.serialization.Serializable
